Improve Player API
Improve Player API so that the ruleset gets passed in during
construction, not at choice retrieval. This is much easier to
use (especially in asynchronous contexts) as it allows the
ruleset to be retrieved prior to the get_choice call.

This also means that players cannot be reused (i.e., if you wanted
to change the game to Rock-Paper-Scissors-Lizard-Spock, the
players would have to be regenerated). This is probably for
the best as enforcing a reusability invariant will increase
code complexity in exchange for minimal functionality.